feat: rename to SettingsStore
